Update badger to a newer commit position.
The folks at dgraph don't like the requirement to update import paths when a
module API changes, so there has been some back-and-forth on the state of the
release. This update was handled by dropping the previous v2 modules release
and explicitly running `go get github.com/dgraph-io/badger@v2`.

The only local changes are to the handling of options.