Keep track of the number of keys.
Out of the box, BadgerDB does not provide a running total of keys.  This is
inconvenient, as it takes a long time to scan a large database to get the
count. Keep track of the count in a file stashed in the database root.  This
presumes the database will not be concurrently modified by another process (or
another thread of the same process).